Continue module decomposition: extract list write operations
